refactor: cleanup format mappings and types with finite instances
There are a few ways to represent mapping data from fultimator or the pdf's format into project-fu's format.  Standardize on a Record<..,..> form.  This gives the benefits of an enum, but better error messages if the original format type changes.

Standardize on using a list as const for types with a finite number of inhabitants.  Enums could have also worked, but come with additional properties that are not desired or used.  For example, iterating over the inhabitants of a Enum requires more work with Object.values and a filter on string vs number.

Additionally cleanup a few other type choices.  Try and use type over enum and interface where possible.