fix: items are not appearing in Project FU 2.4.0-alphas
Project FU 2.4.0 alphas have added a compendium. The folder creation in fu-parser was
not filtering by the folder type, so when it looked up the "Equipment" folder it was
finding the compendium folder.  This made it so that the items would not show up in
any of the sidebars.

This patch filters the folder type so that items will search for an 'Item' folder of
the correct name, and similarly actors will search for an 'Actor' folder.